% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/within.R
\name{within.environment}
\alias{within.environment}
\alias{within.character}
\title{\code{within} methods for \R environments}
\usage{
\method{within}{environment}(data, expr, ...)

\method{within}{character}(data, expr, ..., .pos = 2L)
}
\arguments{
\item{data}{An \R environment, or the name of a (potentially new) attached
environment.}

\item{expr}{The \R expression to evaluate. Automatically quoted.}

\item{...}{Ignored. Added for compatibility with the S3 generic. Throws an
error if any arguments are passed to \code{...}}

\item{.pos}{If creating a new attached environment, the position of the newly
attached environment.}
}
\value{
the \R environment, invisibly
}
\description{
\code{within} methods for \R environments
}
